School/CSP Final/requirements.txt
pydub
matplotlib
numpy
Pillow
PathLib